dnf p5

+++ modules/solr.sh
@@ -4,17 +4,20 @@ source ./lib
 
 function configure_package()
 {
-    show_message "\tCopying service config file..."
-    cp etc/systemd/system/solr.service /etc/systemd/system/ >> /tmp/install.log 2>&1
-    show_result $?
+    if [ $OS == "rhel" ]
+    then
+        show_message "\tCopying service config file..."
+        cp etc/systemd/system/solr.service /etc/systemd/system/ >> /tmp/install.log 2>&1
+        show_result $?
 
-    show_message "\tCopying start file..."
-    cp etc/solr-service.sh /opt/solr/example/ >> /tmp/install.log 2>&1
-    show_result $?
+        show_message "\tCopying start file..."
+        cp etc/solr-service.sh /opt/solr/example/ >> /tmp/install.log 2>&1
+        show_result $?
 
-    show_message "\tChanging start file permissions..."
-    chmod 755 /opt/solr/example/solr-service.sh >> /tmp/install.log 2>&1
-    show_result $?
+        show_message "\tChanging start file permissions..."
+        chmod 755 /opt/solr/example/solr-service.sh >> /tmp/install.log 2>&1
+        show_result $?
+    fi
 
     show_message "\tEnabling solr in firewall..."
     firewall-cmd --permanent --zone=public --add-port=8983/tcp >> /tmp/install.log 2>&1
@@ -24,31 +27,41 @@ function configure_package()
     systemctl restart firewalld.service >> /tmp/install.log 2>&1
     show_result $?
 
-    show_message "\tRestarting solr..."
-    systemctl restart solr.service >> /tmp/install.log 2>&1
-    show_result $?
+    if [ $OS == "rhel" ]
+    then
+        show_message "\tRestarting solr..."
+        systemctl restart solr.service >> /tmp/install.log 2>&1
+        show_result $?
 
-    show_message "\tEnabling solr..."
-    systemctl enable solr.service >> /tmp/install.log 2>&1
-    show_result $?
+        show_message "\tEnabling solr..."
+        systemctl enable solr.service >> /tmp/install.log 2>&1
+        show_result $?
+    fi
 }
 
 function install_package()
 {
-    show_message "Extracting solr..."
-    tar zxvf /tmp/solr.tgz -C /tmp/ >> /tmp/install.log 2>&1
-    show_result $?
-
-    if [ -e /opt/solr/ ]
+    if [ $OS == "fedora" ]
     then
-        show_message "\tRemoving /opt/solr..."
-        rm -rf /opt/solr/ >> /tmp/install.log 2>&1
+        dnf install --assumeyes solr >> /tmp/install.log 2>&1
+        show_result $?
+    elif [ $OS == "rhel" ]
+    then
+        show_message "Extracting solr..."
+        tar zxvf /tmp/solr.tgz -C /tmp/ >> /tmp/install.log 2>&1
+        show_result $?
+
+        if [ -e /opt/solr/ ]
+        then
+            show_message "\tRemoving /opt/solr..."
+            rm -rf /opt/solr/ >> /tmp/install.log 2>&1
+            show_result $?
+        fi
+
+        show_message "Moving solr..."
+        mv /tmp/solr-4.10.2/ /opt/solr >> /tmp/install.log 2>&1
         show_result $?
     fi
-
-    show_message "Moving solr..."
-    mv /tmp/solr-4.10.2/ /opt/solr >> /tmp/install.log 2>&1
-    show_result $?
 }
 
 function download_package()
